requirements.txt
flask
ping3
requests
python-dotenv
pymysql
